requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
mypy==0.770
mypy-extensions==0.4.3
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
